Upgrade to Python 3.12 / Snakemake 8 (#355)
* Upgrade to Python 3.12 / Snakemake 8; migrate from poetry to uv; improve pyrunner test strategy

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add ruff to dev-dependencies

* Remove redundant import

* Revert docs build to Python 3.11

* Update test-repo to use grapevne-helper

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Remove absolute paths from test workflows

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add setuptools to github actions

* Remove absolute path from pyinstaller test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Migrate install_miniforge script to .venv from poetry

* Separate file and folder copying when packaging modules

* Modify Dockerfile for Snakemake v8 and install grapevne-py

* Fix bug that prevented email alerts from being sent

* Remove sources from docs build

* Setup conda using github action for postbuild tests

* Switch payload-shell to payload-run for Windows testing

* Disable workflow checks on Windows where modules connect together (known bug)

* Update nightly and publish workflows to support action-based conda install

* Update nightly and publish workflows to support action-based conda install

* Reject non-native binaries introduced by PuLP

* Remove .venv from build

* Revert and update nightly_build github action

* Disable colima/docker install on macos-12 as both very slow and inconsistent

* Migrate to grapevne-py installer

* Update grapevne_helper in tests

---------

+++ builder/builder/builder.py
@@ -407,19 +407,24 @@ class Model:
         src = os.path.normpath(pathlib.Path(m_path, os.pardir))
         pathlib.Path(dest).mkdir(parents=True, exist_ok=True)
         # Copy module to the build directory
-        ignore_in_root = ["results", "logs", "benchmarks"]
+        ignore_in_root = ["results", "logs", "benchmarks", ".test.sh"]
         ignore_anywhere = [".snakemake", "__pycache__"]
         folders_in_root = os.listdir(src)
         keep_folders = set(folders_in_root) - set(ignore_in_root)
         for folder in keep_folders:
-            shutil.copytree(
-                pathlib.Path(src, folder),
-                pathlib.Path(dest, folder),
-                dirs_exist_ok=True,
-                ignore=lambda directory, contents: contents
-                if any(map(directory.endswith, ignore_anywhere))
-                else set(),
-            )
+            src_folder = pathlib.Path(src, folder)
+            dest_folder = pathlib.Path(dest, folder)
+            if src_folder.is_file():
+                shutil.copy2(src_folder, dest_folder)
+            else:
+                shutil.copytree(
+                    src_folder,
+                    dest_folder,
+                    dirs_exist_ok=True,
+                    ignore=lambda directory, contents: contents
+                    if any(map(directory.endswith, ignore_anywhere))
+                    else set(),
+                )
         # Redirect snakefile location in config
         node.snakefile = str(
             pathlib.Path(
